Some other pretty cool hooks.
