+++ src/screens/AssetsScreen/Upper/Upper.tsx
@@ -1,43 +1,12 @@
-/* eslint-disable @typescript-eslint/no-shadow */
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { View } from 'react-native';
-import { useSharedValue } from 'react-native-reanimated';
-import { useVector } from 'react-native-redash';
-import { getTokenHistory } from '@models/token';
 import Chart from './Chart/Chart';
 import Header from './Header/Header';
-import { GraphIndex, UpperProps } from './Upper.types';
+import { UpperProps } from './Upper.types';
+import { useUpper } from './Upper.hooks';
 
 export const Upper: React.FC<UpperProps> = ({ coin }) => {
-	const [tokenHistory, setTokenHistory] = useState<any>(null);
-	const translation = useVector();
-	const transition = useSharedValue(0);
-	const previous = useSharedValue<GraphIndex>(0);
-	const current = useSharedValue<GraphIndex>(0);
-
-	const fetchTokenHistory = async () => {
-		const res = await getTokenHistory(coin.id);
-		if (res.errors) {
-			const res = await getTokenHistory(coin.name);
-			if (res.errors) {
-				const res = await getTokenHistory(coin.symbol);
-				if (res.errors) {
-					setTokenHistory(null);
-				} else {
-					setTokenHistory(res);
-				}
-			} else {
-				setTokenHistory(res);
-			}
-		} else {
-			setTokenHistory(res);
-		}
-	};
-
-	useEffect(() => {
-		fetchTokenHistory();
-	}, []);
-
+	const { tokenHistory, transition, previous, current, translation } = useUpper(coin);
 	return (
 		<>
 			<Header {...{ coin }} />
